pyproject.toml
[project]
name = "my-doctr-ocr"
version = "0.3.0"
description = "training doctr ocr model"
authors = [
    {name = "dowon", email = "ks2515@naver.com"},
]
dependencies = [
    "torch @ https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp310-cp310-win_amd64.whl ; sys_platform == \"win32\" and python_version == \"3.10\"",
    "torch; sys_platform != 'win32'",
    "torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp310-cp310-win_amd64.whl ; sys_platform == \"win32\" and python_version == \"3.10\"",
    "torchvision; sys_platform != 'win32'",
    "python-doctr @ git+https://github.com/Bing-su/doctr@ko",
    "pytorch-lightning",
    "wandb",
    "pandas",
    "polars",
    "rich",
    "loguru",
    "typer",
    "opencv-python",
    "albumentations",
    "timm",
    "pyyaml",
    "munch",
    "fastprogress>=0.1.21",
    "psutil>=5.9.0",

]
requires-python = "==3.10.*"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black",
    "isort",
    "ruff",
    "mypy",
    "ipywidgets",
]

[tool.pdm.scripts]
isort = "isort ."
black = "black ."
style = {composite = ["isort", "black"]}
ruff = "ruff ."
rufffix = "ruff --fix ."

[tool.isort]
profile = "black"

[tool.ruff]
select = ["A", "B", "C4", "E", "F", "N", "PT", "UP", "W"]
ignore = ["B008", "B905", "E501", "N812"]
unfixable = ["F401"]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
